backend/requirements.txt
Flask==3.1.0
Flask-CORS==5.0.0
PyJWT==2.10.1
Werkzeug==3.1.3
PyPDF2==3.0.1
python-docx==1.1.2
beautifulsoup4==4.12.3
requests==2.32.3
lxml==5.3.0
Flask-Mail==0.10.0
APScheduler==3.10.4
selenium==4.15.2
pyotp==2.9.0
qrcode==7.4.2
Flask-Limiter==3.5.0
gunicorn==21.2.0
Authlib==1.3.0
flask-swagger-ui==4.11.1
flasgger==0.9.7.1
python-dotenv==1.0.0
Pillow==10.1.0
psycopg2-binary
